Nota Dinas

+++ database/migrations/2025_02_06_113135_create_anggarans_table.php
@@ -14,11 +14,11 @@ return new class extends Migration
         Schema::create('anggaran', function (Blueprint $table) {
             $table->id();
             $table->year('tahun');
-            $table->unsignedBigInteger('sub_kegiatan_id');
+            $table->unsignedSmallInteger('sub_kegiatan_id');
             $table->foreign('sub_kegiatan_id')->references('id')->on('ref_kegiatan_sub')->onUpdate('cascade')->onDelete('cascade');
-            $table->unsignedBigInteger('jenis_sppd_id');
+            $table->unsignedTinyInteger('jenis_sppd_id');
             $table->foreign('jenis_sppd_id')->references('id')->on('ref_jenis_sppd')->onUpdate('cascade')->onDelete('cascade');
-            $table->unsignedBigInteger('bidang_sub_id');
+            $table->unsignedSmallInteger('bidang_sub_id');
             $table->foreign('bidang_sub_id')->references('id')->on('ref_bidang_sub')->onUpdate('cascade')->onDelete('cascade');
             $table->double('rp_pagu');
             $table->timestamps();
